♻ refactor: ProductToggleButton 훅 분리
