+++ src/refactoring/pages/admin/components/ProductToggleButton.tsx
@@ -1,7 +1,7 @@
-import { useState } from "react";
-import { Discount, Product } from "../../../../types";
-import UpdateProductForm from "./UpdateProductForm";
+import { Product } from "../../../../types";
+import { useProductToggleButton } from "../hooks/useProductToggleButton";
 import ProductDetails from "./ProductDetails";
+import UpdateProductForm from "./UpdateProductForm";
 
 interface Props {
   products: Product[];
@@ -9,68 +9,18 @@ interface Props {
 }
 
 const ProductToggleButton = ({ products, onProductUpdate }: Props) => {
-  const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const [newDiscount, setNewDiscount] = useState<Discount>({
-    quantity: 0,
-    rate: 0,
-  });
-
-  const toggleProductAccordion = (productId: string) => {
-    setOpenProductIds((prev) => {
-      const newSet = new Set(prev);
-      if (newSet.has(productId)) {
-        newSet.delete(productId);
-      } else {
-        newSet.add(productId);
-      }
-      return newSet;
-    });
-  };
-
-  // handleEditProduct 함수 수정
-  const handleEditProduct = (product: Product) => {
-    setEditingProduct({ ...product });
-  };
-
-  const updateProductField = (field: keyof Product, value: any) => {
-    if (editingProduct) {
-      setEditingProduct({ ...editingProduct, [field]: value });
-    }
-  };
-
-  // 수정 완료 핸들러 함수 추가
-  const handleEditComplete = () => {
-    if (editingProduct) {
-      onProductUpdate(editingProduct);
-      setEditingProduct(null);
-    }
-  };
-
-  const handleAddDiscount = (productId: string) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct && editingProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: [...updatedProduct.discounts, newDiscount],
-      };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-      setNewDiscount({ quantity: 0, rate: 0 });
-    }
-  };
-
-  const handleRemoveDiscount = (productId: string, index: number) => {
-    const updatedProduct = products.find((p) => p.id === productId);
-    if (updatedProduct) {
-      const newProduct = {
-        ...updatedProduct,
-        discounts: updatedProduct.discounts.filter((_, i) => i !== index),
-      };
-      onProductUpdate(newProduct);
-      setEditingProduct(newProduct);
-    }
-  };
+  const {
+    openProductIds,
+    editingProduct,
+    newDiscount,
+    setNewDiscount,
+    toggleProductAccordion,
+    handleEditProduct,
+    updateProductField,
+    handleEditComplete,
+    handleAddDiscount,
+    handleRemoveDiscount,
+  } = useProductToggleButton(products, onProductUpdate);
 
   return (
     <div className="space-y-2">
